Сделано добавление пользователя в бд после регистрации, но без роли, не работает миграция
