+++ back/SCT.Users/DTOs/usersDTO.cs
@@ -6,5 +6,6 @@ namespace SCT.Users.DTOs
         public string? Surname {get; set;}
         public string? Email { get; set; }
         public required string Password {get; set;}
+        //public required string Role { get; set; }
     }
 }
